Feature: Added a computer player and a user player component.
