+++ src/components/GameBoard/GameBoard.js
@@ -1,44 +1,15 @@
-import React, { useEffect, useState } from 'react';
-import Player from '../Player/Player';
+import React from 'react';
+import ComputerPlayer from '../Player/ComputerPlayer';
+import UserPlayer from '../Player/UserPlayer';
 
 function GameBoard() {
-
- // STATE
- let [userPlayer, setUserPlayer] = useState('');
- let [computerPlayer, setComputerPlayer] = useState();
-
- // PROPERTIES
- const computerPlayerNames = {
-  1: 'Dr.Robot_Nick',
-  2: '010110',
-  3: 'Process Sir',
-  4: 'Reboot'
- };
-
- // METHODS
- let addNewComputerPlayer = () => {
-  let computerPlayerValues = Object.values(computerPlayerNames);
-  setComputerPlayer(computerPlayerValues[Math.floor(Math.random() * 3)]);
- };
-
- let addNewUserPlayer = () => {
-  let newPlayer = prompt();
-  setUserPlayer(newPlayer);
- };
-
-
- // useEffect Hook
- useEffect(() => {
-  addNewUserPlayer();
-  addNewComputerPlayer();
- }, []);
 
  return <div className="gameboard">
   {/* Return the two players of the game. */}
   {/* Computer Player */}
-  <Player playerType={computerPlayer} />
+  <ComputerPlayer />
   {/* User Player */}
-  < Player playerType={userPlayer} />
+  <UserPlayer />
  </div>;
 }
