+++ src/components/Player/Player.js
@@ -2,8 +2,14 @@ import React from 'react';
 
 function Player({ playerType }) {
 
+ // Create an HP (Health point) stat
+ let hp = 2000;
+ // Import the Deck
+
+
  return <div className='player'>
-  {playerType}
+  <h1>{playerType}</h1>
+  <span className='hp'>{hp}</span>
  </div>;
 }
